Minor cleanup in thin3d

+++ ext/native/gfx_es2/draw_buffer.cpp
@@ -400,7 +400,7 @@ void DrawBuffer::MeasureTextCount(int font, const char *text, int count, float *
 }
 
 void DrawBuffer::MeasureTextRect(int font, const char *text, int count, const Bounds &bounds, float *w, float *h, int align) {
-	if (!text || (uint32_t)font >= atlas->num_fonts) {
+	if (!text || (uint32_t)font >= (uint32_t)atlas->num_fonts) {
 		*w = 0;
 		*h = 0;
 		return;
